:sparkless: Adicionando barra de acessibilidade e configurando o footer
