+++ SeniorStockManagerFrontend/src/components/Footer/index.tsx
